Removed stuff specific to project 1 game. Updated save file format. Minor modification to Camera to prevent it from going off map. Changed Spawner to specify number of objects to spawn.
Right now, to spawn the player, have a Spawner in the level with properties: obj=Player, rate=0
Later Outhouse and other Player spawners will be added.

Apologies for removing anything that might have been used in any of your maps.  Please tell me about anything that might now be broken so we can try to fix it.

git-svn-id: file:///var/vcs/repos/svn/trunk@33 4d56c4e7-3306-40bb-bc25-9f1c267d86e0