Dialog class for loading and displaying dialog boxes. DialogState has been updated to use it.
See /data/dialogs/example.json for an example of how to create a dialog file. (the only required parameter are the lines)

tool_dialog.py is a tool for testing dialog boxes.  Just drag and drop the file you want to test onto the script then hit any key to progress through the dialog boxes.

git-svn-id: file:///var/vcs/repos/svn/trunk@35 4d56c4e7-3306-40bb-bc25-9f1c267d86e0