Created external tileset for maps.  This should make it easier to create multiple maps using the same tileset as well as allowing changes to the tileset be applied to all maps that use it.
To use an external tileset in Tiled, use Map -> Add External Tileset.

To modify an existing map to use an external tileset, change the tileset tag to <tileset firstgid="1" source="tileset.tsx"/> and remove the stuff inside the tags (along with the closing tag).

To modify the external tileset, open tileset.tmx, make changes, then click the Export Tileset As button in the tileset frame.  Then reload the map you were working on.

git-svn-id: file:///var/vcs/repos/svn/trunk@41 4d56c4e7-3306-40bb-bc25-9f1c267d86e0