Considering whether or not it would be easier to animate 3D characters -- especially since I didn't want to have to redraw the animations for the new resolution.  The major problem would be that only the player really needs a lot of animations, so it might take more time to model, rig, and animate everything else (like enemies, NPCs, etc.)  If we had more time available this might not be a problem.
Another issue is the matter of consistency.  3D graphics for characters standing next to hand drawn tiles and backgrounds might not look right.  Another problem is that anyone who doesn't know how to use Blender cannot easily contribute towards creating graphics.

Anyhow, here is a Blender file for a sheriff character and an image showing roughly the size he would be as a sprite.

git-svn-id: file:///var/vcs/repos/svn/trunk@60 4d56c4e7-3306-40bb-bc25-9f1c267d86e0